+++ src/art.h
@@ -1,5 +1,6 @@
 #ifndef ART_H
 #define ART_H
+#include <cstdint>
 #include <glm/glm.hpp>
 
 #define SCREEN_WIDTH    1024
@@ -8,8 +9,8 @@
 void init_screen(const char * caption);
 void flip_screen();
 
-void pix(int64_t x, int64_t y, int c);
-int rgb(int r, int g, int b);
-int rgb(float r, float g, float b);
+void pix(uint32_t x, uint32_t y, uint32_t c);
+uint32_t rgb(uint32_t r, uint32_t g, uint32_t b);
+uint32_t rgb(float r, float g, float b);
     
 #endif
